% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_players.R
\name{load_players}
\alias{load_players}
\title{Load RFL players via mfl api}
\usage{
load_players()
}
\arguments{
\item{season:}{A numeric vector of 4-digit years}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{player_name} (character)
\item \code{team} (character)
\item \code{pos} (character)
\item \code{player_id} (integer)
\item \code{status} (character)
}
}
\description{
Loads league players via mfl players API endpoint
}
\examples{
players <- rflfastR::load_players()

head(players, 10)
}
