% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_playerscores.R
\name{load_playerscores}
\alias{load_playerscores}
\title{Load RFL player score data from github}
\usage{
load_playerscores(seasons = 2016:nflreadr::most_recent_season())
}
\arguments{
\item{season:}{A numeric vector of 4-digit years}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{season} (integer)
\item \code{week} (integer)
\item \code{player_id} (integer)
\item \code{player_name} (character)
\item \code{pos} (character)
\item \code{team} (character)
\item \code{points} (double)
}
}
\description{
Loads rfl-playerscores.rds files from github into a dataframe
}
\examples{
starter <- rflfastR::load_playerscores(2022)

head(starter, 10)
}
