% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_starter.R
\name{load_starter}
\alias{load_starter}
\title{Load RFL starter data from github}
\usage{
load_starter(seasons = 2016:nflreadr::most_recent_season())
}
\arguments{
\item{season:}{A numeric vector of 4-digit years}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{season} (integer)
\item \code{week} (integer)
\item \code{franchise_id} (character)
\item \code{starter_status} (character)
\item \code{player_id} (integer)
\item \code{player_name} (character)
\item \code{pos} (character)
\item \code{team} (character)
\item \code{player_score} (double)
\item \code{should_start} (integer)
}
}
\description{
Loads rfl-starter.csv files from github into a dataframe
}
\examples{
starter <- rflfastR::load_starter(2022)

head(starter, 10)
}
